Move the event bus worker test to EventBusTestBase to execute them also in local mode
